fixed overflow bug

+++ tests/test_lib.rs
@@ -1,4 +1,4 @@
-use nalgebra::{ComplexField, Vector2, Vector4};
+use nalgebra::{Vector2, Vector4};
 use snake_rust::Snake;
 
 #[test]
@@ -11,29 +11,29 @@ fn test_snake() {
     assert!(snake.body[0].x == 1 && snake.body[0].y == 1);
     assert!(snake.direction.x == 0 && snake.direction.y == 0);
 
-    let mut direction = Vector2::<i8>::new(1, 0);
+    let mut direction = Vector2::<i16>::new(1, 0);
     snake.set_direction(&direction);
     snake.step();
     assert!(snake.body[0].x == 3 && snake.body[0].y == 1);
     assert!(snake.direction.x == 1 && snake.direction.y == 0);
 
-    let food = Vector2::<i8>::new(3, 1);
+    let food = Vector2::<i16>::new(3, 1);
     assert!(snake.try_eat_food(&food) == true);
     assert!(snake.body.len() == 2);
     assert!(snake.body[1].x == 3 && snake.body[1].y == 1);
 
-    direction = Vector2::<i8>::new(-1, 0);
+    direction = Vector2::<i16>::new(-1, 0);
     snake.set_direction(&direction);
     assert!(snake.direction.x == 1 && snake.direction.y == 0);
 
-    direction = Vector2::<i8>::new(0, 0);
+    direction = Vector2::<i16>::new(0, 0);
     snake.set_direction(&direction);
     assert!(snake.direction.x == 1 && snake.direction.y == 0);
 
     snake.speed_up();
     assert!((snake.speed - 1.2).abs() < 0.1);
 
-    let boundaries = Vector4::<i8>::new(0, 0, 1, 1);
+    let boundaries = Vector4::<i16>::new(0, 0, 1, 1);
 
     assert!(snake.try_hit_walls(&boundaries) == true)
 }
